bpf: l2_responder: split map declarations into l2_responder_maps.h
l2_responder.h mixes the L2 responder map types/declarations (plain
data, includable from any BPF program) with handle_l2_announcement(),
which pulls in <bpf/config/global.h> (interface_mac/ifindex configs),
ARP and ICMPv6 helpers, and the agent-liveness runtime config. A BPF
program that only wants to LOOK UP the responder maps (to ask "is this
address one this node currently L2-announces?") is forced to satisfy
the full announcement-handler include chain, which is not available
outside bpf_host.

Move the map types/declarations into a new l2_responder_maps.h and
include it from l2_responder.h, so existing consumers are unchanged.
The enable_l2_announcements config declaration moves here too: a
consumer deciding whether to consult the maps needs the gate, and it
depends only on the lightweight static_data.h rather than the handler
include chain.

No behavior change.
